[WIP] Trying to split js_panel_window into parts

+++ foo_spider_monkey_panel/js_objects/window.h
@@ -51,7 +51,7 @@ public: // methods
 public: // props
     std::optional<uint32_t> get_DlgCode();
     std::optional<uint32_t> get_Height();
-    std::optional<uint32_t> get_Id();
+    std::optional<uint64_t> get_Id();
     std::optional<uint32_t> get_InstanceType();
     std::optional<bool> get_IsTransparent();
     std::optional<bool> get_IsVisible();
